refactor invitations
This PR removes the need for an `accepted` field in `TeamInvitationEntity`. When the user registers, the entry of `TeamInvitation` is removed.  Moreover, it renames `TeamService::findPendingInvitations` to `TeamService:: findInvitations`

Creates single exception for the registratin process:

`org.projectcheckins.security.RegistrationCheckViolationException`.

Deletes both `UserAlreadyExistsException`and `UserNotInvitedException`.

The `RegistrationCheck` api now returns an `Optional<RegistrationCheckViolation>` which allows for multiple ordered  registration checks to be added.